Add DoughnutChart component to BalanceBox
This commit adds the DoughnutChart component to the BalanceBox in the balance-box.tsx file. The DoughnutChart component is imported from the 'component' module and is rendered within the total-balance-chart div. This addition enhances the user interface by displaying a visual representation of the account balances in a doughnut chart format.