feat: Add DoughnutChart component
This commit adds the DoughnutChart component to the repository. The component is implemented in the `component/chart/doughnut.tsx` file and provides a visual representation of account balances in a doughnut chart format. This addition enhances the user interface by displaying account data in a more visually appealing and intuitive way.