🐳 chore(feat: Add @radix-ui/react-tabs dependency
This commit adds the @radix-ui/react-tabs dependency to the package.json file. The @radix-ui/react-tabs library provides tabbed navigation functionality, which will be used to enhance the user interface of the application. This addition improves the user experience by allowing users to navigate between different sections of the application using tabs.):