Add new SVG icons for credit card and arrow navigation
This commit adds new SVG icons for credit card and arrow navigation. The icons are located in the asset directory and can be used in various parts of the application. This enhancement improves the user experience by providing visually appealing and intuitive icons for credit card-related actions and arrow navigation.